Bundle SDS ingress tests together
