+++ tests/integration/security/sds_ingress/single_mtls_gateway_separate_secret_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package singlemtlsgatewayseparatesecret
+package sdsingress 
 
 import (
 	"testing"
