+++ tests/integration/security/sds_ingress/single_tls_gateway_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package singletlsgateway
+package sdsingress 
 
 import (
 	"testing"
